Move parse_document code into Document
