+++ src/cobalt.rs
@@ -10,8 +10,7 @@ use walkdir::WalkDir;
 use document::Document;
 use error::{Error, Result};
 use config::Config;
-use yaml_rust::YamlLoader;
-use chrono::{DateTime, UTC, FixedOffset};
+use chrono::{UTC, FixedOffset};
 use chrono::offset::TimeZone;
 use rss::{Channel, Rss};
 use std::sync::Arc;
@@ -71,7 +70,7 @@ pub fn build(config: &Config) -> Result<()> {
                                               .extension()
                                               .unwrap_or(OsStr::new(""))) &&
            entry.path().parent() != Some(layouts_path.as_path()) {
-            let mut doc = try!(parse_document(&entry.path(), &source));
+            let mut doc = try!(Document::parse(&entry.path(), &source));
 
             // if the document is in the posts folder it's considered a post
             if entry.path().parent() == Some(posts_path.as_path()) {
@@ -231,73 +230,3 @@ fn create_rss(path: &str, dest: &Path, config: &Config, documents: &[Document]) 
     }
 }
 
-fn parse_document(path: &Path, source: &Path) -> Result<Document> {
-    let mut attributes = HashMap::new();
-    let mut content = try!(parse_file(path));
-
-    // if there is front matter, split the file and parse it
-    if content.contains("---") {
-        let content2 = content.clone();
-        let mut content_splits = content2.split("---");
-
-        // above the split are the attributes
-        let attribute_string = content_splits.next().unwrap_or("");
-
-        // everything below the split becomes the new content
-        content = content_splits.next().unwrap_or("").to_owned();
-
-        let yaml_result = try!(YamlLoader::load_from_str(attribute_string));
-
-        let yaml_attributes = try!(yaml_result[0]
-                                       .as_hash()
-                                       .ok_or(format!("Incorrect front matter format in {:?}",
-                                                      path)));
-
-        for (key, value) in yaml_attributes {
-            // TODO is unwrap_or the best way to handle this?
-            attributes.insert(key.as_str().unwrap_or("").to_owned(),
-                              value.as_str().unwrap_or("").to_owned());
-        }
-    }
-
-    let date = attributes.get("date")
-                         .and_then(|d| DateTime::parse_from_str(d, "%d %B %Y %H:%M:%S %z").ok());
-
-    let path_str = try!(path.to_str()
-                            .ok_or(format!("Cannot convert pathname {:?} to UTF-8", path)));
-
-    let source_str = try!(source.to_str()
-                                .ok_or(format!("Cannot convert pathname {:?} to UTF-8", source)));
-
-    let new_path = try!(path_str.split(source_str)
-                                .last()
-                                .ok_or(format!("Empty path")));
-
-    // construct path
-    let mut path_buf = PathBuf::from(new_path);
-    path_buf.set_extension("html");
-
-    let path_str = try!(path_buf.to_str()
-                                .ok_or(format!("Cannot convert pathname {:?} to UTF-8", path_str)));
-
-    let markdown = path.extension().unwrap_or(OsStr::new("")) == OsStr::new("md");
-
-    let name = try!(path.file_stem()
-                        .and_then(|stem| stem.to_str())
-                        .ok_or(format!("Invalid UTF-8 in file stem for {:?}", path)));
-
-    Ok(Document::new(name.to_owned(),
-                     path_str.to_owned(),
-                     attributes,
-                     content,
-                     false,
-                     date,
-                     markdown))
-}
-
-fn parse_file(path: &Path) -> Result<String> {
-    let mut file = try!(File::open(path));
-    let mut text = String::new();
-    try!(file.read_to_string(&mut text));
-    Ok(text)
-}
